@@ -4,6 +4,7 @@ import (
 	"encoding/xml"
 	"fmt"
 	"mynews/internal/pkg/timeparser"
+	"time"
 )
 
 type atomFeed struct {
@@ -21,22 +22,29 @@ type atomLink struct {
 	Href string `xml:"href,attr"`
 }
 
-func parseAtom(body []byte) (items []Item, err error) {
-	var r atomFeed
+func parseAtom(body []byte) ([]Item, error) {
+	var feed atomFeed
 
-	if err = xml.Unmarshal(body, &r); err != nil {
+	if err := xml.Unmarshal(body, &feed); err != nil {
 		return nil, fmt.Errorf("failed to parse Atom feed: %w", err)
 	}
 
-	for _, feedItem := range r.Items {
-		item := Item{Title: feedItem.Title, Link: feedItem.Link.Href, PublishedAt: feedItem.Updated}
+	items := make([]Item, len(feed.Items))
 
-		item.PublishedAtParsed, err = timeparser.ParseUTC(feedItem.Updated)
+	for itemIdx := range feed.Items {
+		items[itemIdx] = Item{
+			Title:             feed.Items[itemIdx].Title,
+			Link:              feed.Items[itemIdx].Link.Href,
+			PublishedAt:       feed.Items[itemIdx].Updated,
+			PublishedAtParsed: time.Time{},
+		}
+
+		publishedAtParsed, err := timeparser.ParseUTC(feed.Items[itemIdx].Updated)
 		if err != nil {
 			return nil, fmt.Errorf("failed to parse feed item publish date: %w", err)
 		}
 
-		items = append(items, item)
+		items[itemIdx].PublishedAtParsed = publishedAtParsed
 	}
 
 	return items, nil
